+++ example/lib/tabs/overflow_day_view_tab.dart
@@ -1,5 +1,6 @@
+import 'dart:collection';
+
 import 'package:calendar_day_view/calendar_day_view.dart';
-import 'package:example/utils.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 
@@ -7,28 +8,30 @@ class OverflowDayViewTab extends HookWidget {
   const OverflowDayViewTab({
     Key? key,
     required this.events,
+    this.onTimeTap,
   }) : super(key: key);
   final List<DayEvent<String>> events;
-
+  final Function(TimeOfDay)? onTimeTap;
   @override
   Widget build(BuildContext context) {
     final timeGap = useState<int>(60);
+    final colorScheme = Theme.of(context).colorScheme;
     return Column(
       children: [
         Expanded(
           child: OverFlowCalendarDayView(
-            onTimeTap: print,
-            events: events,
+            onTimeTap: onTimeTap ?? print,
+            events: UnmodifiableListView(events),
             dividerColor: Colors.black,
             startOfDay: const TimeOfDay(hour: 00, minute: 0),
             endOfDay: const TimeOfDay(hour: 23, minute: 0),
             timeGap: timeGap.value,
+            heightPerMin: .4,
             renderRowAsListView: true,
             showCurrentTimeLine: true,
             showMoreOnRowButton: true,
-            overflowItemBuilder: (context, constraints, event) {
+            overflowItemBuilder: (context, constraints, itemIndex, event) {
               return HookBuilder(builder: (context) {
-                final randomColor = useMemoized(() => getRandomColor());
                 return GestureDetector(
                   key: ValueKey(event.hashCode),
                   onTap: () {
@@ -36,6 +39,7 @@ class OverflowDayViewTab extends HookWidget {
                   },
                   child: Container(
                     margin: const EdgeInsets.only(right: 3),
+                    padding: const EdgeInsets.symmetric(horizontal: 5),
                     key: ValueKey(event.hashCode),
                     width: constraints.minWidth < 100
                         ? 100
@@ -44,12 +48,15 @@ class OverflowDayViewTab extends HookWidget {
                         constraints.minWidth - 3,
                     height: constraints.maxHeight,
                     decoration: BoxDecoration(
-                      color: randomColor,
-                      borderRadius: const BorderRadius.all(Radius.circular(10)),
+                      color: itemIndex % 2 == 0
+                          ? colorScheme.primary
+                          : colorScheme.secondary,
                     ),
                     child: Center(
                       child: Text(
+                        textAlign: TextAlign.center,
                         event.value,
+                        overflow: TextOverflow.fade,
                         style: const TextStyle(
                           color: Colors.white,
                           fontWeight: FontWeight.bold,