+++ lib/calendar_day_view.dart
@@ -9,4 +9,7 @@ library calendar_day_view;
 export './src/day_views/event_calendar_day_view.dart';
 export './src/day_views/in_row_calendar_day_view.dart';
 export './src/day_views/overflow_calendar_day_view.dart';
+export './src/day_views/category_calendar_day_view.dart';
 export 'src/models/day_event.dart';
+export 'src/models/categorized_day_event.dart';
+export 'src/models/event_category.dart';