+++ lib/src/day_views/overflow_calendar_day_view.dart
@@ -95,9 +95,8 @@ class OverFlowCalendarDayView<T extends Object> extends StatefulWidget {
 
 class _OverFlowCalendarDayViewState<T extends Object>
     extends State<OverFlowCalendarDayView<T>> {
-  List<TimeOfDay> _timesInDay = [];
   List<OverflowEventsRow<T>> _overflowEvents = [];
-  double _heightPerMin = 1.0;
+
   TimeOfDay _currentTime = TimeOfDay.now();
   Timer? _timer;
   double _rowScale = 1;
@@ -107,11 +106,7 @@ class _OverFlowCalendarDayViewState<T extends Object>
     super.initState();
     _rowScale = 1;
 
-    _heightPerMin = widget.heightPerMin;
-    _timesInDay = getTimeList();
-
-    _overflowEvents = processOverflowEvents(widget.events
-      ..sort(
+    _overflowEvents = processOverflowEvents([...widget.events]..sort(
         (a, b) => a.compare(b),
       ));
 
@@ -127,13 +122,9 @@ class _OverFlowCalendarDayViewState<T extends Object>
   @override
   void didUpdateWidget(covariant OverFlowCalendarDayView<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-
-    _timesInDay = getTimeList();
-    _overflowEvents = processOverflowEvents(widget.events
-      ..sort(
+    _overflowEvents = processOverflowEvents([...widget.events]..sort(
         (a, b) => a.compare(b),
       ));
-    _heightPerMin = widget.heightPerMin;
   }
 
   @override
@@ -148,7 +139,6 @@ class _OverFlowCalendarDayViewState<T extends Object>
     final timeCount =
         (((timeEnd.hour + 1) - widget.startOfDay.hour) * 60) ~/ widget.timeGap -
             1;
-
     DateTime first = DateTime.parse(
         "2012-02-27T${widget.startOfDay.hour.toString().padLeft(2, '0')}:00");
 
@@ -162,8 +152,10 @@ class _OverFlowCalendarDayViewState<T extends Object>
 
   @override
   Widget build(BuildContext context) {
-    final heightUnit = _heightPerMin * _rowScale;
-    final rowHeight = widget.timeGap * _heightPerMin * _rowScale;
+    final heightUnit = widget.heightPerMin * _rowScale;
+    final rowHeight = widget.timeGap * widget.heightPerMin * _rowScale;
+
+    final timesInDay = getTimeList();
 
     return LayoutBuilder(builder: (context, constraints) {
       final viewWidth = constraints.maxWidth;
@@ -176,16 +168,16 @@ class _OverFlowCalendarDayViewState<T extends Object>
           physics: widget.physics,
           padding: const EdgeInsets.only(top: 10, bottom: 20),
           child: SizedBox(
-            height: _timesInDay.length * rowHeight,
+            height: timesInDay.length * rowHeight,
             child: Stack(
               clipBehavior: Clip.none,
               // padding: const EdgeInsets.only(top: 20, bottom: 20),
               children: [
                 ListView.builder(
                   physics: const NeverScrollableScrollPhysics(),
-                  itemCount: _timesInDay.length,
+                  itemCount: timesInDay.length,
                   itemBuilder: (context, index) {
-                    final time = _timesInDay.elementAt(index);
+                    final time = timesInDay.elementAt(index);
                     return GestureDetector(
                       key: ValueKey(time.toString()),
                       behavior: HitTestBehavior.opaque,
@@ -236,17 +228,6 @@ class _OverFlowCalendarDayViewState<T extends Object>
                     width: constraints.maxWidth,
                     color: widget.currentTimeLineColor,
                   ),
-                // GestureDetector(
-                //   onScaleUpdate: (details) {
-                //     if (details.horizontalScale != _rowScale) {
-                //       setState(() {
-                //         if (details.scale >= 1 && details.scale <= 10) {
-                //           _rowScale = details.scale;
-                //         }
-                //       });
-                //     }
-                //   },
-                // ),
               ],
             ),
           ),
@@ -258,10 +239,18 @@ class _OverFlowCalendarDayViewState<T extends Object>
   List<Widget> renderWithFixedWidth(double heightUnit, double eventColumnWith) {
     return [
       for (final oEvents in _overflowEvents)
-        for (final event in oEvents.events)
+        for (var i = 0; i < oEvents.events.length; i++)
           Builder(
             builder: (context) {
+              final event = oEvents.events.elementAt(i);
               final width = eventColumnWith / oEvents.events.length;
+              final tileConstraints = BoxConstraints(
+                maxHeight: event.durationInMins * heightUnit,
+                minHeight: event.durationInMins * heightUnit,
+                minWidth: width,
+                maxWidth: eventColumnWith,
+              );
+
               return Positioned(
                 left: widget.timeTitleColumnWidth +
                     oEvents.events.indexOf(event) * width,
@@ -269,12 +258,8 @@ class _OverFlowCalendarDayViewState<T extends Object>
                     heightUnit,
                 child: widget.overflowItemBuilder!(
                   context,
-                  BoxConstraints(
-                    maxHeight: event.durationInMins * heightUnit,
-                    minHeight: event.durationInMins * heightUnit,
-                    minWidth: width,
-                    maxWidth: eventColumnWith,
-                  ),
+                  tileConstraints,
+                  i,
                   event,
                 ),
               );
@@ -283,6 +268,7 @@ class _OverFlowCalendarDayViewState<T extends Object>
     ];
   }
 
+// to render all events in same row as a horizontal istView
   List<Widget> renderAsListView(double heightUnit, double eventColumnWith) {
     return [
       for (final oEvents in _overflowEvents)