+++ lib/src/extensions/list_extensions.dart
@@ -1,0 +1,10 @@
+
+extension ListX<T> on Iterable<T> {
+  T? firstWhereOrNull(bool Function(T element) test) {
+    final l = where(test);
+    if (l.isNotEmpty) {
+      return l.first;
+    }
+    return null;
+  }
+}