+++ lib/src/models/event_category.dart
@@ -1,0 +1,21 @@
+class EventCategory {
+  final String id;
+  final String name;
+  EventCategory({
+    required this.id,
+    required this.name,
+  });
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+
+    return other is EventCategory && other.id == id && other.name == name;
+  }
+
+  @override
+  int get hashCode => id.hashCode ^ name.hashCode;
+
+  @override
+  String toString() => 'EventCategory(id: $id, name: $name)';
+}