+++ lib/src/widgets/overflow_list_view_row.dart
@@ -33,11 +33,12 @@ class OverflowListViewRow<T extends Object> extends StatefulWidget {
 class _OverflowListViewRowState<T extends Object>
     extends State<OverflowListViewRow<T>> {
   late ScrollController _scrollCtrl;
-  bool _atEndOfList = false;
+  bool _atEndOfList = true;
 
   @override
   void initState() {
     super.initState();
+    _atEndOfList = true;
     _scrollCtrl = ScrollController();
 
     _scrollCtrl.addListener(() {
@@ -94,6 +95,14 @@ class _OverflowListViewRowState<T extends Object>
               final event = widget.oEvents.events.elementAt(index);
               final width =
                   widget.eventColumnWith / widget.oEvents.events.length;
+
+              final tileConstraints = BoxConstraints(
+                maxHeight: event.durationInMins * widget.heightUnit,
+                minHeight: event.durationInMins * widget.heightUnit,
+                minWidth: width,
+                maxWidth: widget.eventColumnWith,
+              );
+
               return Column(
                 children: [
                   SizedBox(
@@ -108,12 +117,8 @@ class _OverflowListViewRowState<T extends Object>
                       ignored: widget.ignored,
                       child: widget.overflowItemBuilder(
                         context,
-                        BoxConstraints(
-                          maxHeight: event.durationInMins * widget.heightUnit,
-                          minHeight: event.durationInMins * widget.heightUnit,
-                          minWidth: width,
-                          maxWidth: widget.eventColumnWith,
-                        ),
+                        tileConstraints,
+                        index,
                         event,
                       ),
                     ),